Add files via upload
Verifying that the first, second and third derivatives of x-, y- and z-coordinates calculating from Nagy et al. (2010) are equal to Sharma (1956)
In Nagy et al. (2000) R = $ [ (xo - x) ² + (yo - y)² + (zo - z)² ]^1/2 $, where x, y and z are the coordinates of the data and xo, yo and zo are the positions of the mass
In Sharma (1956) R_1 = $ [ (x  - xo) ² + (y  - yo)² + (  z - c)² ]^1/2 $ where x, y and z are the coordinates of the data and xo, yo and c are the positions of the nucleus 

The derivatives are identical. 

References
Nagy, D., Papp, G. & Benedek, J. The gravitational potential and its derivatives for the prism. Journal of Geodesy 74, 552–560 (2000). 
B.D. Sharma, Stresses in an infinite slab due to a nucleus of thermoelastic strain in it, Z. Angew. Math. Mech. 36 (1–2) (1956) 565–589.